Feat/#93 커뮤니티 게시판 타입으로 리스트 조회

+++ community-service/src/main/java/com/freediving/communityservice/application/service/ArticleService.java
@@ -2,12 +2,19 @@ package com.freediving.communityservice.application.service;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import com.freediving.communityservice.adapter.in.web.UserProvider;
+import com.freediving.communityservice.adapter.out.dto.article.ArticleBriefDto;
 import com.freediving.communityservice.adapter.out.dto.article.ArticleContentWithComment;
 import com.freediving.communityservice.application.port.in.ArticleEditCommand;
+import com.freediving.communityservice.application.port.in.ArticleIndexListCommand;
 import com.freediving.communityservice.application.port.in.ArticleReadCommand;
 import com.freediving.communityservice.application.port.in.ArticleRemoveCommand;
 import com.freediving.communityservice.application.port.in.ArticleUseCase;
@@ -40,19 +47,20 @@ public class ArticleService implements ArticleUseCase {
 	//Query
 	@Override
 	public Article getArticle(ArticleReadCommand command) {
-		return articleReadPort.readArticle(command.getBoardId(), command.getArticleId(), command.isShowAll());
+		return articleReadPort.readArticle(command.getBoardType(), command.getArticleId(), command.isShowAll());
 	}
 
 	@Override
 	public ArticleContentWithComment getArticleWithComment(ArticleReadCommand command) {
 
 		if (command.isWithoutComment()) {
-			Article onlyArticle = articleReadPort.readArticle(command.getBoardId(), command.getArticleId(),
+			Article onlyArticle = articleReadPort.readArticle(command.getBoardType(), command.getArticleId(),
 				command.isShowAll());
 			return new ArticleContentWithComment(onlyArticle, null);
 		}
 
-		ArticleContentWithComment foundContent = articleReadPort.readArticleWithComment(command);
+		ArticleContentWithComment foundContent = articleReadPort.readArticleWithComment(command.getBoardType(),
+			command.getArticleId(), false);
 		//TODO 글 작성자에게만 보여지는 값 추가시 사용 Long articleOwner = articleContentWithComment.getArticle().getCreatedBy();
 
 		UserProvider requestUser = command.getUserProvider();
@@ -64,12 +72,36 @@ public class ArticleService implements ArticleUseCase {
 		return new ArticleContentWithComment(foundContent.getArticle(), filteredComments);
 	}
 
+	@Override
+	public Page<ArticleBriefDto> getArticleIndexList(ArticleIndexListCommand command) {
+		Pageable pageable = PageRequest.of(command.getPage(), command.getOffset(),
+			Sort.by(Sort.Direction.DESC, command.getOrderBy()));
+		Page<ArticleBriefDto> pagingArticleList = articleReadPort.retrieveArticleIndexList(command.getBoardType(),
+			command.getCursor(), pageable);
+
+		//TODO ArticleBriefDto pagingArticleList.getContent() 멤버 서비스 : 사용자 상태 및 profileImg 요청
+		//TODO "pageable"-"pageNumber": 1 등의 값만 사용. 응답값 재구성
+
+		List<Long> articleOwnerIds = pagingArticleList.getContent().stream()
+			.map(ArticleBriefDto::getCreatedBy)
+			.toList();
+
+		/*
+		 * memberClient.getMemberInfo( articleOwnerIds );
+		 *
+		 * */
+
+		return pagingArticleList;
+
+	}
+
 	//Command
 	@Override
 	public Long writeArticle(ArticleWriteCommand articleWriteCommand) {
 		//TODO 한 사용자는 같은 게시판에 00초 내로 글 추가가 제한됨. 프론트에서 먼저 체크 등.
 		// articleReadPort.getXXXByAuthorId()
-		Board board = boardReadPort.findById(articleWriteCommand.getBoardId());
+		Optional<Board> foundBoard = boardReadPort.findByBoardType(articleWriteCommand.getBoardType());
+		Board board = foundBoard.orElseThrow(() -> new IllegalArgumentException("해당하는 게시판이 없습니다."));
 		board.checkPermission(articleWriteCommand);
 		Article savedArticle = articleWritePort.writeArticle(articleWriteCommand);
 		//TODO articleWriteCommand Hashtag 저장
@@ -79,10 +111,11 @@ public class ArticleService implements ArticleUseCase {
 
 	@Override
 	public Long editArticle(ArticleEditCommand command) {
-		Article originalArticle = articleReadPort.readArticle(command.getBoardId(), command.getArticleId(), false);
+		Article originalArticle = articleReadPort.readArticle(command.getBoardType(), command.getArticleId(), false);
 		originalArticle.checkHasOwnership(command.getUserProvider().getRequestUserId());
 
-		Long updatedArticleId = articleEditPort.updateArticle(command.getBoardId(), command.getArticleId(), command.getTitle(), command.getContent(), command.getHashtagIds(), command.isEnableComment());
+		Long updatedArticleId = articleEditPort.updateArticle(command.getBoardType(), command.getArticleId(),
+			command.getTitle(), command.getContent(), command.getHashtagIds(), command.isEnableComment());
 
 		return null;
 	}
@@ -92,10 +125,12 @@ public class ArticleService implements ArticleUseCase {
 		// TODO Admin 등 처리?
 		boolean isShowAll = false;
 
-		Article article = articleReadPort.readArticle(command.getBoardId(), command.getArticleId(), isShowAll);
+		Article article = articleReadPort.readArticle(command.getBoardType(), command.getArticleId(), isShowAll);
 		if (Objects.isNull(article)) {
 			throw new IllegalArgumentException("해당하는 게시글이 없습니다.");
 		}
+		article.checkHasOwnership(command.getUserProvider().getRequestUserId());
+
 		commentDeletePort.deleteComments(article.getId());
 
 		return articleDeletePort.removeArticle(command);
